+++ scripts/cmd/start.sh
@@ -23,7 +23,7 @@ start_core_services() {
     if [ "$all_docker" = true ]; then
         docker-compose up -d
     else
-        docker-compose up -d postgres redis mailpit apisix envoy rabbitmq kafka-cmd
+        docker-compose up -d postgres redis mailpit apisix rabbitmq kafka-cmd
         # Ensure Docker versions of application services are stopped to prevent port conflicts (native execution preferred)
         docker stop gridtokenx-trading-service gridtokenx-iam-service gridtokenx-aggregator-bridge gridtokenx-noti-service gridtokenx-chain-bridge >/dev/null 2>&1 || true
     fi